requirements.txt
altgraph==0.16.1
astroid==2.3.3
certifi==2019.11.28
cffi==1.13.2
colorama==0.4.1
cryptography==2.8
future==0.18.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.17.4
pefile==2019.4.18
pycparser==2.19
PyInstaller==3.5
pylab==0.0.2
pylint==2.4.3
pynput==1.5.2
python-dotenv==0.10.3
python-telegram-bot==12.2.0
pyusb==1.0.2
PyVISA==1.10.1
PyVISA-py==0.3.1
pywin32-ctypes==0.2.0
six==1.13.0
tornado==6.0.3
wrapt==1.11.2
